site/requirements.txt
beautifulsoup4==4.5.1
bleach==1.5.0
boto3==1.9.23
botocore==1.12.23
certifi==2018.8.24
chardet==2.3.0
defusedxml==0.5.0
Django==1.11.15
django-contrib-comments==1.7.3
django-debug-toolbar==1.10.1
django-game-info==0.1.3
django-haystack==2.5.0
django-nocaptcha-recaptcha==0.0.19
django-pagination-py3==1.1.2
django-storages==1.7.1
djangorestframework==3.4.6
docutils==0.14
filebrowser-safe==0.4.5
future==0.15.2
grappelli-safe==0.4.4
gunicorn==19.9.0
html5lib==0.9999999
jmespath==0.9.3
Markdown==2.6.6
mysqlclient==1.3.13
oauthlib==2.0.0
Pillow==3.3.1
postmarkup==1.2.2
Pygments==2.1.3
PyJWT==1.5.3
python-dateutil==2.7.3
python-openid==2.2.5
python-social-auth==0.2.21
python-valve==0.1.1
python3-openid==3.1.0
pytz==2016.6.1
requests==2.11.1
requests-oauthlib==0.6.2
s3transfer==0.1.13
sentry-sdk==0.4.0
six==1.10.0
sqlparse==0.2.4
tzlocal==1.2.2
urllib3==1.23
